requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
requests
asyncio
toml
eventlet
gunicorn
aiofiles==24.1.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.0
cachelib==0.13.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
dnspython==2.6.1
dpath==2.2.0
email_validator==2.2.0
eventlet==0.33.3
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
Flask==2.3.3
Flask-Limiter==3.10.1
Flask-Login==0.6.2
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
html5tagger==1.3.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.2
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
mysql-connector==2.2.9
numpy==2.0.1
opencv-python==4.10.0.84
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PySocks==1.7.1
pyTBA==1.2.1
pyTelegramBotAPI==4.23.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-multipart==0.0.10
python-slugify==8.0.4
python-socketio==5.13.0
python-telegram-bot==21.4
pyTelegramBotAPI==4.23.0
pytz==2025.2
requests==2.32.4
rich==13.9.4
sanic==24.6.0
sanic-routing==23.12.0
schedule==1.2.2
selenium==4.13.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.21
starlette==0.38.5
telebot==0.0.5
toml==0.10.2
tracerite==1.1.1
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.7
wrapt==1.17.2
wsproto==1.2.0
betfairlightweight==2.23.2
WTForms==3.2.1
yarl==1.13.1
zipp==3.21.0
